Add files via upload
This the Capstone project where K-Means clustering was used alongside Foursquare API and public economic data to identify clusters of financial cities based on a given set of features. I believe this project to be successful given that the output resembled the perceived risks of the financial cities analysed in this.